Update the package name
