+++ src/main/java/com/healthai/springboot/demo/edi/DentalEDIApplication.java
@@ -1,4 +1,4 @@
-package com.luv2code.springboot.demo.edi;
+package com.healthai.springboot.demo.edi;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
