+++ src/main/java/com/healthai/springboot/demo/edi/model/Claim.java
@@ -1,0 +1,5 @@
+package com.healthai.springboot.demo.edi.model;
+
+
+public class Claim {
+}
