+++ src/main/java/com/healthai/springboot/demo/edi/model/Recipient.java
@@ -1,0 +1,4 @@
+package com.healthai.springboot.demo.edi.model;
+
+public class Recipient {
+}
