+++ src/main/java/com/healthai/springboot/demo/edi/repository/ClaimRepository.java
@@ -1,0 +1,4 @@
+package com.healthai.springboot.demo.edi.repository;
+
+public class ClaimRepository {
+}
