+++ src/main/java/com/healthai/springboot/demo/edi/service/ClaimService.java
@@ -1,0 +1,4 @@
+package com.healthai.springboot.demo.edi.service;
+
+public class ClaimService {
+}
