--- src/main/java/com/luv2code/springboot/demo/edi/model/Claim.java
@@ -1,5 +1,0 @@
-package com.luv2code.springboot.demo.edi.model;
-
-@Entity
-public class Claim {
-}
