--- src/main/java/com/luv2code/springboot/demo/edi/model/Recipient.java
@@ -1,4 +1,0 @@
-package com.luv2code.springboot.demo.edi.model;
-
-public class Recipient {
-}
