--- src/main/java/com/luv2code/springboot/demo/edi/repository/ClaimRepository.java
@@ -1,4 +1,0 @@
-package com.luv2code.springboot.demo.edi.repository;
-
-public class ClaimRepository {
-}
