--- src/main/java/com/luv2code/springboot/demo/edi/service/ClaimService.java
@@ -1,4 +1,0 @@
-package com.luv2code.springboot.demo.edi.service;
-
-public class ClaimService {
-}
