+++ src/test/java/com/healthai/springboot/demo/edi/DentalEDIApplicationTests.java
@@ -1,4 +1,4 @@
-package com.luv2code.springboot.demo.edi;
+package com.healthai.springboot.demo.edi;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
